requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
async-lru==2.0.5
axios==0.1.0
beanie==2.0.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.0.1
click==8.1.8
click-spinner==0.1.10
contourpy==1.3.2
croniter==6.0.0
cryptography==46.0.2
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
esprima==4.0.1
fastapi==0.115.12
fastapi-cli==0.0.7
fonttools==4.57.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.1
h11==0.16.
httpcore==1.0.9
httptools==0.6.4
httpx==1.0.dev2
idna==3.4
importlib-metadata==6.0.0
isodate==0.7.2
Jinja2==3.1.4
kiwisolver==1.4.7
lazy-model==0.3.0
librecaptcha==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
meteomatics==2.11.3
motor==3.7.1
narwhals==2.7.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
openpyxl==3.1.5
passlib==1.7.4
bcrypt==4.1.3
pathlib2==2.3.7.post1
pillow==11.0.0
plotly==5.24.1
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.6.1
pydantic_core==2.33.1
Pygments==2.18.0
PyJWT==2.10.1
PyKrige==1.7.2
pymongo==4.15.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.19
pytz==2025.2
PyWeatherKit @ git+https://github.com/joaoTrevizoli/PyWeatherKit.git@c10a2961d8b8fe87e4cdbbd9ef65bb0eeba91f41
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
toml==0.10.2
truststore==0.10.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.3
weatherkit==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
